run_demo.py
#!/usr/bin/env python3
"""
Script de demonstração do classificador de áudio
Este script executa uma versão simplificada para demonstrar o funcionamento
"""

import os
import numpy as np
import torch
import torch.nn as nn
from torchvision import models, transforms
from PIL import Image
import matplotlib.pyplot as plt

def create_demo_data():
    """Cria dados de demonstração para testar o sistema"""
    print("🔧 Criando dados de demonstração...")
    
    # Criar diretórios
    os.makedirs('demo_data/mel_spectrograms/class_a', exist_ok=True)
    os.makedirs('demo_data/mel_spectrograms/class_b', exist_ok=True)
    os.makedirs('demo_data/stft_spectrograms/class_a', exist_ok=True)
    os.makedirs('demo_data/stft_spectrograms/class_b', exist_ok=True)
    
    # Criar imagens dummy (simulando espectrogramas)
    for i in range(10):
        # Criar imagem aleatória 224x224x3
        dummy_image_data = np.random.randint(0, 255, size=(224, 224, 3), dtype=np.uint8)
        dummy_image = Image.fromarray(dummy_image_data)
        
        # Salvar imagens para cada classe e tipo
        dummy_image.save(f'demo_data/mel_spectrograms/class_a/mel_a_{i}.png')
        dummy_image.save(f'demo_data/mel_spectrograms/class_b/mel_b_{i}.png')
        dummy_image.save(f'demo_data/stft_spectrograms/class_a/stft_a_{i}.png')
        dummy_image.save(f'demo_data/stft_spectrograms/class_b/stft_b_{i}.png')
    
    print("✅ Dados de demonstração criados!")

def load_resnet_model(num_classes):
    """Carrega modelo ResNet18 com Transfer Learning"""
    model = models.resnet18(pretrained=True)
    num_ftrs = model.fc.in_features
    model.fc = nn.Linear(num_ftrs, num_classes)
    return model

def demo_training():
    """Demonstra o processo de treinamento (simulado)"""
    print("\n🧠 Demonstrando processo de treinamento...")
    
    # Configurações
    num_classes = 2
    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
    print(f"Usando dispositivo: {device}")
    
    # Criar modelos
    mel_model = load_resnet_model(num_classes)
    stft_model = load_resnet_model(num_classes)
    
    mel_model.to(device)
    stft_model.to(device)
    
    # Simular treinamento (apenas para demonstração)
    print("📊 Simulando treinamento do modelo Mel-espectrograma...")
    print("   Epoch 1/2 - Loss: 0.6931 - Acc: 0.5000")
    print("   Epoch 2/2 - Loss: 0.4521 - Acc: 0.7500")
    
    print("📊 Simulando treinamento do modelo STFT-espectrograma...")
    print("   Epoch 1/2 - Loss: 0.6845 - Acc: 0.5250")
    print("   Epoch 2/2 - Loss: 0.4123 - Acc: 0.8000")
    
    # Salvar modelos (apenas os pesos iniciais para demonstração)
    torch.save(mel_model.state_dict(), 'demo_mel_model.pth')
    torch.save(stft_model.state_dict(), 'demo_stft_model.pth')
    
    print("✅ Modelos de demonstração salvos!")
    
    return mel_model, stft_model

def demo_evaluation(mel_model, stft_model):
    """Demonstra a avaliação dos modelos"""
    print("\n📈 Demonstrando avaliação dos modelos...")
    
    # Simular resultados de avaliação
    print("🎧 Mel-espectrograma:")
    print("   Acurácia: 0.7500 (75.00%)")
    print("   Precisão: 0.7600")
    print("   Recall: 0.7400")
    
    print("📊 STFT-espectrograma:")
    print("   Acurácia: 0.8000 (80.00%)")
    print("   Precisão: 0.8100")
    print("   Recall: 0.7900")
    
    print("🤝 Ensemble:")
    print("   Acurácia: 0.8500 (85.00%)")
    print("   Precisão: 0.8600")
    print("   Recall: 0.8400")
    
    print("✅ Avaliação concluída!")

def demo_prediction():
    """Demonstra uma predição em tempo real"""
    print("\n🔍 Demonstrando predição em tempo real...")
    
    # Simular carregamento de áudio
    print("📁 Carregando arquivo de áudio: 'exemplo.wav'")
    print("🎵 Gerando mel-espectrograma...")
    print("📊 Gerando espectrograma STFT...")
    
    # Simular predições
    print("\n🧠 Fazendo predições:")
    print("   Mel-espectrograma: Classe A (confiança: 72.3%)")
    print("   STFT-espectrograma: Classe B (confiança: 68.9%)")
    print("   Ensemble: Classe A (confiança: 78.5%)")
    
    print("✅ Predição concluída!")

def main():
    """Função principal da demonstração"""
    print("🎵 DEMONSTRAÇÃO DO CLASSIFICADOR DE ÁUDIO")
    print("=" * 50)
    
    # Criar dados de demonstração
    create_demo_data()
    
    # Demonstrar treinamento
    mel_model, stft_model = demo_training()
    
    # Demonstrar avaliação
    demo_evaluation(mel_model, stft_model)
    
    # Demonstrar predição
    demo_prediction()
    
    print("\n" + "=" * 50)
    print("🎉 DEMONSTRAÇÃO CONCLUÍDA!")
    print("\nPara usar o sistema completo:")
    print("1. Execute: python3 audio_classifier.py (para treinamento)")
    print("2. Execute: streamlit run real_time_interface.py (para interface)")
    print("\nArquivos criados:")
    print("- audio_classifier.py: Script principal de treinamento")
    print("- real_time_interface.py: Interface web Streamlit")
    print("- requirements.txt: Dependências do projeto")
    print("- README.md: Documentação completa")

if __name__ == "__main__":
    main()

