requirements.txt
torch>=1.9.0
torchvision>=0.10.0
librosa>=0.8.1
matplotlib>=3.3.4
numpy>=1.21.0
Pillow>=8.3.0
streamlit>=1.10.0
scikit-learn>=1.0.0
soundfile>=0.10.3

